return of hardware_usart.cpp it contains one function, namely set_peripheral because it causes linking errors somehow
